+++ src/pages/MyPage/pages/DeleteAccountPage/DeleteAccountPage.jsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 
 import { useAuth, useModal } from '@/hooks';
 
-import { CloseAppBar, ConfirmModal, InputPassword } from '@/shared/component';
+import { CloseAppBar, ConfirmModal, PwInput } from '@/shared/component';
 
 import styles from './DeleteAccountPage.module.css';
 
@@ -44,7 +44,7 @@ export default function DeleteAccountPage() {
             ))}
           </div>
 
-          <InputPassword
+          <PwInput
             title='비밀번호'
             placeholder='비밀번호를 입력하세요'
             value={password}
